BACKLOG-21819: unselect nodes when changing folder
