+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentTable/ContentTable.jsx
@@ -1,4 +1,4 @@
-import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import React, {useCallback, useMemo, useRef, useState} from 'react';
 import clsx from 'clsx';
 import PropTypes from 'prop-types';
 import {registry} from '@jahia/ui-extender';
@@ -14,7 +14,6 @@ import {extractPaths, getCanDisplayItemParams} from '~/JContent/JContent.utils';
 import {shallowEqual, useDispatch, useSelector} from 'react-redux';
 import {cmSetPreviewSelection} from '~/JContent/redux/preview.redux';
 import {cmSetPage, cmSetPageSize} from '~/JContent/redux/pagination.redux';
-import {cmRemoveSelection} from '~/JContent/redux/selection.redux';
 import JContentConstants from '~/JContent/JContent.constants';
 import ContentEmptyDropZone from './ContentEmptyDropZone';
 import ContentNotFound from './ContentNotFound';
@@ -31,13 +30,10 @@ import {cmSetSort} from '~/JContent/redux/sort.redux';
 import {Row} from '~/JContent/ContentRoute/ContentLayout/ContentTable/Row';
 import {useFileDrop} from '~/JContent/dnd/useFileDrop';
 import styles from './ContentTable.scss';
-import {pathExistsInTree} from '../../../JContent.utils';
-import {useNotifications} from '@jahia/react-material';
-import {TableViewModeChangeTracker} from '../../ToolBar/ViewModeSelector/tableViewChangeTracker';
+import {useUnselect} from '~/JContent/ContentRoute/ContentLayout/useUnselect';
 
 export const ContentTable = ({rows, isContentNotFound, totalCount, isLoading, isStructured, columns, selector}) => {
     const {t} = useTranslation('jcontent');
-    const {notify} = useNotifications();
     const dispatch = useDispatch();
 
     const {mode, previewSelection, siteKey, path, pagination, previewState, selection, searchTerms, tableOpenPaths, sort} = useSelector(selector, shallowEqual);
@@ -89,45 +85,7 @@ export const ContentTable = ({rows, isContentNotFound, totalCount, isLoading, is
         useExpandedControlled
     );
 
-    useEffect(() => {
-        if (selection.length > 0 && !isLoading) {
-            const notVisible = (rows?.length > 0) ? selection.filter(path => !pathExistsInTree(path, rows)) : selection;
-            if (notVisible.length > 0) {
-                const toRemove = [];
-                notVisible.forEach(currentPath => {
-                    const toOpen = [];
-                    if (isStructured && currentPath.startsWith(path)) {
-                        let pathParts = currentPath.substring(path.length).split('/').slice(0, -1);
-                        let pathToAdd = '';
-                        for (let pathPart of pathParts) {
-                            pathToAdd = pathToAdd ? (pathToAdd + '/' + pathPart) : path;
-                            if (tableOpenPaths.indexOf(pathToAdd) === -1) {
-                                toOpen.push(pathToAdd);
-                            }
-                        }
-                    }
-
-                    if (toOpen.length === 0) {
-                        // The node was not visible, and we cannot fix that by opening folders: remove selection
-                        toRemove.push(currentPath);
-                    } else {
-                        dispatch(cmOpenTablePaths([...new Set(toOpen)]));
-                    }
-                });
-
-                if (toRemove.length > 0) {
-                    dispatch(cmRemoveSelection(toRemove));
-                    if (TableViewModeChangeTracker.modeChanged) {
-                        notify(t('jcontent:label.contentManager.selection.removed', {count: toRemove.length}), ['closeButton', 'closeAfter5s']);
-                    }
-                }
-            }
-
-            TableViewModeChangeTracker.resetChanged();
-        } else if (!isLoading && rows?.length > 0) {
-            TableViewModeChangeTracker.resetChanged();
-        }
-    }, [rows, tableOpenPaths, selection, dispatch, path, paths, isLoading, notify, isStructured, t]);
+    useUnselect(selection, isLoading, rows, isStructured, path, tableOpenPaths);
 
     const doubleClickNavigation = useCallback(node => {
         let newMode = mode;
